+++ darglint/driver.py
@@ -197,7 +197,7 @@ def print_error_list():
 
 
 def print_version():
-    print('1.2.3')
+    print('1.2.4')
 
 
 def main():