+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.4]
+
+### Fixed
+
+- Handle newline after exception in raises section.  Rather
+  than failing to parse the raises section, it now reports
+  an indentation error.
+
+
 ## [1.2.3]
 
 ### Changed